+++ tests/test_processor.py
@@ -6,23 +6,16 @@ import time
 import typing
 import unittest.mock
 
-import asynctest
-
 from sprockets_statsd import statsd
 from tests import helpers
 
 
-class ProcessorTestCase(asynctest.TestCase):
+class ProcessorTestCase(unittest.IsolatedAsyncioTestCase):
     ip_protocol: int
 
-    async def setUp(self):
+    def setUp(self):
         self.test_timeout = 5.0
         super().setUp()
-        await self.asyncSetUp()
-
-    async def tearDown(self):
-        await self.asyncTearDown()
-        super().tearDown()
 
     async def wait_for(self, fut):
         try:
@@ -31,6 +24,7 @@ class ProcessorTestCase(asynctest.TestCase):
             self.fail('future took too long to resolve')
 
     async def asyncSetUp(self):
+        await super().asyncSetUp()
         self.statsd_server = helpers.StatsdServer(self.ip_protocol)
         self.statsd_task = asyncio.create_task(self.statsd_server.run())
         await self.statsd_server.wait_running()
@@ -38,6 +32,7 @@ class ProcessorTestCase(asynctest.TestCase):
     async def asyncTearDown(self):
         self.statsd_server.close()
         await self.statsd_server.wait_closed()
+        await super().asyncTearDown()
 
 
 class ProcessorTests(ProcessorTestCase):