+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
-Next release
-------------
+:tag:`1.0.0 <0.1.0...1.0.0>` (20-Jul-2021)
+------------------------------------------
 - Added ``Connector.timer`` method (addresses :issue:`8`)
 - Implement ``Timer`` abstraction from python-statsd
 - ``Connector.timing`` now accepts :class:`datetime.timedelta` instances in addition